requirements_dataset_testing.txt
pandas>=1.3.0
numpy>=1.21.0
Pillow>=8.3.0
rasterio>=1.2.0
